Create AGFS and LGFS namespace for fee type rules
They AGFS and LGFS require different rules in
some cases. Share common functionality in concern.
