+++ app/validators/fee/base_fee_validator.rb
@@ -93,7 +93,7 @@ module Fee
     end
 
     def validate_fee_type_rules
-      rule_sets = Fee::FeeTypeRules.where(unique_code: @record.fee_type&.unique_code)
+      rule_sets = Fee::AGFS::FeeTypeRules.where(unique_code: @record.fee_type&.unique_code)
       fee_type_rule_validator = Rule::Validator.new(@record, rule_sets)
       fee_type_rule_validator.validate
     end
