+++ spec/support/shared_examples/shared_examples_for_fee_type_rules.rb
@@ -1,5 +1,7 @@
-RSpec.describe Fee::FeeTypeRules, type: :validator do
-  it { is_expected.to respond_to(:sets) }
+# frozen_string_literal: true
+
+RSpec.shared_examples 'fee_type_rules_creator' do
+   it { is_expected.to respond_to(:sets) }
 
   before do
     create(:misc_fee_type, :miumu)
